+++ R/bins.R
@@ -6,8 +6,12 @@
 
 #' Sinusoidal coordinates from bin number
 #' 
-#' bin L3 bin number
-#' NUMROWS grid spec for these bins
+#'
+#' 
+#'
+#' @param bin  bin L3 bin number
+#' @param NUMROWS NUMROWS grid spec for these bins
+#'
 #' @export
 sinu_coords <- function(bin, NUMROWS) {
     rgdal::project(do.call(cbind, bin2lonlat(bin, NUMROWS)), "+proj=sinu +datum=WGS84 +no_defs")