+++ R/regrid.R
@@ -4,7 +4,7 @@
   dims <- c(4320L, 720L)
   sensor <- match.arg(sensor)
   if(sensor == "SeaWiFS") dims <- dims/2L
-  setValues(raster(extent(-180, 180, -90, -30), nrows = dims[2L], ncols = dims[1L], crs = "+proj=longlat +datum=WGS84"), 
+  raster::setValues(raster(extent(-180, 180, -90, -30), nrows = dims[2L], ncols = dims[1L], crs = "+proj=longlat +datum=WGS84"), 
             rep(0, prod(dims)))
 }
 
@@ -32,50 +32,7 @@ lonlatFromCell <- function(object, cell = NULL, spatial = FALSE) {
   }
 }
 
-##' Regrid ocean colour. 
-##' 
-##' Regrid NASA ocean colour RRS values to standard Mapped image with the Johnson chl-a algorithm. 
-##'
-##' @param tgrid target grid
-##' @param file L3 bin file name with raw RRS wavelengths
-##' @param agg aggregate values?
-##' @param NROWS number of L3 rows
-##' @param fun function to aggregate by if agg is \code{TRUE}
-##'
-##' @importFrom rgdal project
-##' @importFrom raster setValues
-##' @export
-regridder <- function(tgrid, file, agg = FALSE, fun = mean, NROWS) {
-  xy <- lonlatFromCell(tgrid, spatial = FALSE)
-  
-  stop("need to use read_l3_file somehow")   #MDS 2017-05-25
-  #NROWS <- readL3(file, vname = "NUMROWS", bins = FALSE)$NUMROWS
-  binmap <- lonlat2bin(xy[,1], xy[,2], NROWS)
-  function(file) {
-    ## stop MDS 2017-05-25
-    #x <- readL3(file)
-    sens <- if(x$NUMROWS == 4320) "MODISA" else "SeaWiFS"
-    if(x$NUMROWS != NROWS)  stop("file doesn't match this regridder\nfile NUMROWS: ", x$NUMROWS, "\nfunction NUMROWS:", NROWS)
-   
-    if (agg) {
-     vals <-  chla(x, sensor = sens, algo = "johnson")
-     
-     binxy <- do.call(cbind, bin2lonlat(x$bin_num, x$NUMROWS))
-     if (!is.na(projection(tgrid)) & !isLonLat(tgrid)) binxy <- project(binxy, projection(tgrid))
-     gcell <- extract(tgrid, binxy, cellnumbers = TRUE)[,"cells"]
-     tvals <- tapply(vals, gcell, fun)
-     
-     r <- tgrid
-     r[unique(gcell)] <- tvals
-     r[!r>0] <- NA
-    } else {
-      r <- setValues(tgrid, chla(x, sensor = sens, algo = "johnson")[match(binmap, x$bin_num)])
-    }
-    r
-  }
-  
-  
-}
+
 
 
 # start <- 120